third_party/blink/web_tests/external/wpt/mathml/presentation-markup/operators/mo-font-relative-lengths-001-expected.txt

This is a testharness.js-based test.
PASS font-relative lspace refers to the core operator
PASS font-relative rspace refers to the core operator
FAIL font-relative minsize refers to the core operator assert_approx_equals: baseSizePx * fontScalePercent * minsizeEm expected 400 +/- 1 but got 50
PASS font-relative maxsize refers to the core operator
Harness: the test ran to completion.
